tests/wpt/web-platform-tests/content-security-policy/style-src/style-src-3_4.html

<!doctype html>
<html>
<head>
    <title></title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link href="style-src-3_4.css?pipe=sub" rel=stylesheet type=text/css>
</head>
<body>
    <h1>
        @import stylesheet should not be loaded
        if its URL doesn't match style-src.
    </h1>
    <div id='log'></div>

    <script>
    test(function() {
        var logEl = document.getElementById("log");
        var marginLeftVal = getComputedStyle(logEl).getPropertyValue('margin-left');
        assert_false(marginLeftVal == "200px")},
        "@import stylesheet should not be applied"
    );
    </script>

    <script async defer src='../support/checkReport.sub.js?reportField=violated-directive&reportValue=style-src%20%27self%27'></script>
</body>
</html>
